remove: thanks count from torrent icons
Significantly cleans up the listing and isn't needed to see in bulk. Some sites disable thanks completely and don't want to see it. I plan to also move the comment count and the seeding/leeching indicators elsewhere in a separate PR to clean it up further. The count is still shown on the individual torrent page itself.